requirements.txt
python-telegram-bot == 22.0
torch == 2.6.0
torchvision == 0.21.0
segmentation_models_pytorch == 0.5.0